backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email] # Modificado para incluir la validación de email
pyodbc
python-dotenv
uuid
